Modularize tracking and utility functions
