+++ helpers/__init__.py
@@ -1,4 +1,27 @@
 # Expose helper functions via relative import
+from .track_markers_range import track_markers_range
+from ._update_nf_and_motion_model import _update_nf_and_motion_model
+from .track_full_clip import track_full_clip
+from .run_iteration import run_iteration
+from ._run_test_cycle import _run_test_cycle
+from .run_pattern_size_test import run_pattern_size_test
+from .evaluate_motion_models import evaluate_motion_models
+from .evaluate_channel_combinations import evaluate_channel_combinations
+from .track_bidirectional import track_bidirectional
+from .track_forward_only import track_forward_only
+from .add_timer import add_timer
+from .pattern_base import pattern_base
+from .pattern_limits import pattern_limits
+from .clamp_pattern_size import clamp_pattern_size
+from .strip_prefix import strip_prefix
+from .add_pending_tracks import add_pending_tracks
+from .clean_pending_tracks import clean_pending_tracks
+from .rename_pending_tracks import rename_pending_tracks
+from .update_frame_display import update_frame_display
+from .cycle_motion_model import cycle_motion_model
+from .compute_detection_params import compute_detection_params
+from .detect_new_tracks import detect_new_tracks
+from .remove_close_tracks import remove_close_tracks
 from .utils import *
 from .feature_math import (
     calculate_base_values,
