+++ helpers/pattern_base.py
@@ -1,0 +1,5 @@
+
+def pattern_base(clip):
+    """Return the default pattern size based on the clip width."""
+    width, _ = clip.size
+    return int(width / 100)
